# Project-Mannager-System/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PM Daily Work Management API
  version: 1.0.0
  description: |
    API cho ứng dụng quản lý công việc hằng ngày của quản lý dự án phần mềm.

    ## Quy ước chung
    - Base path: `/api/v1`
    - Xác thực: `Authorization: Bearer <accessToken>` (JWT HS256, hết hạn 15 phút). Refresh token (7 ngày) chỉ gửi ở `/auth/refresh` và `/auth/logout`.
    - Thời gian: ISO-8601 UTC (`timestamptz`). Ngày (date): `yyyy-MM-dd`.
    - Phân trang: `page` (≥ 0, mặc định 0), `size` (1–100, mặc định 20), `sort` (whitelist theo từng endpoint, VD `fullName,asc`).
    - Lỗi: format thống nhất (BR-GEN-05) — `timestamp, status, error, code, message, path, fieldErrors, traceId`. Catalog mã lỗi: `docs/design/05-error-handling-design.md`.
    - Dữ liệu đã xóa mềm: không xuất hiện ở danh sách; truy cập trực tiếp trả 404.
    - Cập nhật bắt buộc kèm `version`; version cũ → 409.
    - Chi tiết từng module: `docs/api/01..12-*.md`.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /projects:
    get:
      tags: [Project]
      summary: Danh sách dự án (search/filter/paging)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ProjectStatus' } }
        - { name: myOnly, in: query, schema: { type: boolean } }
        - { name: projectManagerId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectPage' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
    post:
      tags: [Project]
      summary: Tạo dự án (PM trở thành thành viên)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreateRequest' }
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '409': { $ref: '#/components/responses/Duplicate409' }
  /projects/{id}:
    get:
      tags: [Project]
      summary: Chi tiết dự án
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
      responses:
        '200':
          description: Thông tin dự án
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    put:
      tags: [Project]
      summary: Cập nhật dự án (optimistic locking)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdateRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
        '409': { $ref: '#/components/responses/Conflict409' }
    delete:
      tags: [Project]
      summary: Xóa mềm dự án (confirm khi còn task mở)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { name: confirm, in: query, schema: { type: boolean, default: false } }
      responses:
        '204': { description: Thành công, không body }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /projects/{id}/members:
    get:
      tags: [Project]
      summary: Danh sách thành viên dự án
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
      responses:
        '200':
          description: Danh sách thành viên
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProjectMemberResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    post:
      tags: [Project]
      summary: Thêm thành viên vào dự án
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectMemberRequest' }
      responses:
        '201':
          description: Đã thêm
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectMemberResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
        '409': { $ref: '#/components/responses/Duplicate409' }
  /projects/{id}/members/{userId}:
    put:
      tags: [Project]
      summary: Đổi vai trò thành viên trong dự án
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectMemberRequest' }
      responses:
        '200':
          description: Đã cập nhật vai trò
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectMemberResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    delete:
      tags: [Project]
      summary: Xóa thành viên khỏi dự án (chặn xóa PM cuối)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Thành công, không body }
        '400':
          description: PROJECT_MANAGER_REQUIRED khi xóa PM cuối
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  # ============================= TASK =============================
...
  /reports/tasks-by-status:
    get:
      tags: [Report]
      summary: Báo cáo task theo trạng thái
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdQueryRequired' }
        - { name: fromDate, in: query, schema: { type: string, format: date } }
        - { name: toDate, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Thống kê
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusCountItems' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /reports/tasks-by-assignee:
    get:
      tags: [Report]
      summary: Báo cáo task theo người thực hiện
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdQueryRequired' }
        - { name: fromDate, in: query, schema: { type: string, format: date } }
        - { name: toDate, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Thống kê
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssigneeCountItems' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /reports/overdue-tasks:
    get:
      tags: [Report]
      summary: Báo cáo task quá hạn (chi tiết)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdQueryRequired' }
        - { $ref: '#/components/parameters/Page' }
        - { name: size, in: query, schema: { type: integer, maximum: 100, default: 50 } }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskSummaryPage' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /reports/project-progress:
    get:
      tags: [Report]
      summary: Báo cáo tiến độ dự án
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: projectId, in: query, required: true, schema: { type: array, items: { type: string, format: uuid }, maxItems: 50 } }
      responses:
        '200':
          description: Tiến độ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectProgressReportResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /reports/risk-issue-summary:
    get:
      tags: [Report]
      summary: Tổng hợp risk & issue mở
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdQueryRequired' }
      responses:
        '200':
          description: Tổng hợp
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RiskIssueSummaryResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /reports/export:
    get:
      tags: [Report]
      summary: Xuất file báo cáo (xlsx/csv)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: report, in: query, required: true, schema: { type: string, enum: [tasks-by-status, tasks-by-assignee, overdue-tasks, project-progress, risk-issue-summary] } }
        - { name: format, in: query, schema: { type: string, enum: [xlsx, csv], default: xlsx } }
        - { $ref: '#/components/parameters/ProjectIdQuery' }
        - { name: fromDate, in: query, schema: { type: string, format: date } }
        - { name: toDate, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: File báo cáo
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string, format: binary }
        '400':
          description: EXPORT_LIMIT_EXCEEDED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  # ============================= AUDIT =============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
      description: Số trang (từ 0)
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Số bản ghi mỗi trang
    Sort:
      name: sort
      in: query
      schema: { type: string }
      description: 'Sắp xếp theo whitelist của endpoint, VD `fullName,asc`'
    ProjectIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: ID dự án
    ProjectIdQuery:
      name: projectId
      in: query
      schema: { type: string, format: uuid }
      description: Lọc theo dự án
    ProjectIdQueryRequired:
      name: projectId
      in: query
      required: true
      schema: { type: string, format: uuid }
      description: ID dự án (bắt buộc)
...
  responses:
    Validation400:
      description: Dữ liệu không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Forbidden403:
      description: Không đủ quyền hoặc ngoài phạm vi dữ liệu
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound404:
      description: Không tìm thấy (bao gồm dữ liệu đã xóa mềm)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict409:
      description: Xung đột optimistic locking (version cũ)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Duplicate409:
      description: Dữ liệu trùng (username/email/mã)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [timestamp, status, error, code, message, path, fieldErrors, traceId]
      properties:
        timestamp: { type: string, format: date-time, example: '2026-08-01T10:00:00Z' }
        status: { type: integer, example: 400 }
        error: { type: string, example: BAD_REQUEST }
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: 'Dữ liệu không hợp lệ' }
        path: { type: string, example: /api/v1/tasks }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        traceId: { type: string, example: a1b2c3d4-1234 }
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, hasNext, hasPrevious]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        hasNext: { type: boolean }
        hasPrevious: { type: boolean }
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
    # ---------- Enum ----------
...
    TaskStatus: { type: string, enum: [TODO, IN_PROGRESS, BLOCKED, REVIEW, DONE, CANCELLED] }
    TaskPriority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    TaskType: { type: string, enum: [FEATURE, BUG, IMPROVEMENT, TASK, OTHER] }
    TaskSource: { type: string, enum: [MANUAL, MEETING, ACTION_ITEM, ISSUE, OTHER] }
    ProjectStatus: { type: string, enum: [PLANNING, ACTIVE, ON_HOLD, COMPLETED, CANCELLED] }
    ProjectMemberRole: { type: string, enum: [PROJECT_MANAGER, TECH_LEAD, BUSINESS_ANALYST, DEVELOPER, TESTER, DEVOPS, MEMBER] }
...
    RiskLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
...
    IssueSeverity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
...
    ProjectCreateRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string, minLength: 3, maxLength: 20 }
        name: { type: string, maxLength: 100 }
        description: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        customerName: { type: string, maxLength: 100 }
        projectManagerId: { type: string, format: uuid }
        note: { type: string }
    ProjectUpdateRequest:
      allOf:
        - { $ref: '#/components/schemas/ProjectCreateRequest' }
        - type: object
          required: [version]
          properties:
            version: { type: integer }
    ProjectResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        projectManagerId: { type: string, format: uuid }
        customerName: { type: string }
        progress: { type: integer }
        memberCount: { type: integer }
        createdAt: { type: string, format: date-time }
        version: { type: integer }
    ProjectMemberRequest:
      type: object
      required: [userId, role]
      properties:
        userId: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/ProjectMemberRole' }
    ProjectMemberResponse:
      type: object
      properties:
        userId: { type: string, format: uuid }
        username: { type: string }
        fullName: { type: string }
        email: { type: string }
        role: { $ref: '#/components/schemas/ProjectMemberRole' }
        joinedAt: { type: string, format: date-time }
    ProjectPage:
      allOf:
        - { $ref: '#/components/schemas/PageResponse' }
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ProjectResponse' }
    # ---------- Task ----------
...
    TaskResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PRJ001-TASK-000001 }
        projectId: { type: string, format: uuid }
        projectCode: { type: string }
        projectName: { type: string }
        parentTaskId: { type: string, format: uuid, nullable: true }
        title: { type: string }
        description: { type: string }
        reporter: { $ref: '#/components/schemas/UserRef' }
        assignee: { $ref: '#/components/schemas/UserRef' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        type: { $ref: '#/components/schemas/TaskType' }
        source: { $ref: '#/components/schemas/TaskSource' }
        startDate: { type: string, format: date, nullable: true }
        dueDate: { type: string, format: date, nullable: true }
        actualCompletedAt: { type: string, format: date-time, nullable: true }
        progress: { type: integer }
        blocked: { type: boolean }
        blockerReason: { type: string, nullable: true }
        estimateMinutes: { type: integer, nullable: true }
        actualMinutes: { type: integer, nullable: true }
        notes: { type: string }
        tags:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              name: { type: string }
              color: { type: string }
        collaborators:
          type: array
          items: { $ref: '#/components/schemas/UserRef' }
        watchers:
          type: array
          items: { $ref: '#/components/schemas/UserRef' }
        commentCount: { type: integer }
        attachmentCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer }
    TaskSummaryResponse:
      allOf:
        - { $ref: '#/components/schemas/TaskResponse' }
        - type: object
          description: Rút gọn cho danh sách (bỏ description/notes đầy đủ nếu cần)
...
    TaskSummaryPage:
      allOf:
        - { $ref: '#/components/schemas/PageResponse' }
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/TaskSummaryResponse' }
    # ---------- Meeting ----------
...
    StatusCountItems:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              status: { $ref: '#/components/schemas/TaskStatus' }
              count: { type: integer }
    AssigneeCountItems:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              assigneeId: { type: string, format: uuid }
              fullName: { type: string }
              count: { type: integer }
              doneCount: { type: integer }
    ProjectProgressReportResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              projectId: { type: string, format: uuid }
              code: { type: string }
              name: { type: string }
              progress: { type: integer }
              totalTasks: { type: integer }
              doneTasks: { type: integer }
    RiskIssueSummaryResponse:
      type: object
      properties:
        openRisks: { type: integer }
        openIssues: { type: integer }
        risksByLevel:
          type: array
          items:
            type: object
            properties:
              level: { $ref: '#/components/schemas/RiskLevel' }
              count: { type: integer }
        issuesBySeverity:
          type: array
          items:
            type: object
            properties:
              severity: { $ref: '#/components/schemas/IssueSeverity' }
              count: { type: integer }
...